requirements.txt
customtkinter>=5.0.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
vgamepad>=0.1.0
pyinstaller>=5.0.0